refactor folder structure.
